pyproject.toml
[tool.poetry]
name = "pypolycontain"
version = "1.4.1"
description = "A python package for polytopic objects, operations, and containment encodings"
authors = ["Sadra Sadraddini <sadra@mit.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
Sphinx = "^7.3.7"
ipykernel = "^6.29.4"
nbsphinx = "^0.9.3"
scipy = "^1.13.0"

# Optional
drake = [{version = "<0.1", source = "drake-repo"}]
pycddlib = "^2.1.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "drake-repo"
url = "https://drake-packages.csail.mit.edu/whl/nightly/"
secondary = true